// bao-source/bao-shared/src/generated/flatbuffers/baohaus/realtime/v1/attitude.ts
// automatically generated by the FlatBuffers compiler, do not modify
import { Builder } from '@baohaus/flatbuf-bao/builder';
import type { IGeneratedObject, Offset } from '@baohaus/flatbuf-bao/builder';
import { ByteBuffer } from '@baohaus/flatbuf-bao/reader';
import type { IUnpackableObject } from '@baohaus/flatbuf-bao/reader';
export class Attitude implements IUnpackableObject<AttitudeT> {
  bb: ByteBuffer|null = null;
  bb_pos = 0;
  __init(i:number, bb:ByteBuffer):Attitude {
  this.bb_pos = i;
  this.bb = bb;
  return this;
}

roll():number {
  return this.bb!.readFloat32(this.bb_pos);
}

pitch():number {
  return this.bb!.readFloat32(this.bb_pos + 4);
}

yaw():number {
  return this.bb!.readFloat32(this.bb_pos + 8);
}

static sizeOf():number {
  return 12;
}

static createAttitude(builder:Builder, roll: number, pitch: number, yaw: number):Offset {
  builder.prep(4, 12);
  builder.writeFloat32(yaw);
  builder.writeFloat32(pitch);
  builder.writeFloat32(roll);
  return builder.offset();
}


unpack(): AttitudeT {
  return new AttitudeT(
    this.roll(),
    this.pitch(),
    this.yaw()
  );
}


unpackTo(_o: AttitudeT): void {
  _o.roll = this.roll();
  _o.pitch = this.pitch();
  _o.yaw = this.yaw();
}
}

export class AttitudeT implements IGeneratedObject {
constructor(
  public roll: number = 0.0,
  public pitch: number = 0.0,
  public yaw: number = 0.0
){}


pack(builder:Builder): Offset {
  return Attitude.createAttitude(builder,
    this.roll,
    this.pitch,
    this.yaw
  );
}
}